fix(db): a COUNT() is INT8, and a TIMESTAMPTZ read into a String
Two read-side twins of the INT4 class, both found by the suite rather than by
the checker, and both in statements that were otherwise correct:

  governance.rs  -- a COUNT() column read as i32. PostgreSQL counts in INT8.
  media_resilience.rs -- media_references.first_seen_at is TIMESTAMPTZ and the
    row reader asks for a String, so the column needs ::text on output.

Also repairs bounties.rs, where an earlier line-index edit of mine had dropped a
closing quote and left a stray `use serde_json::Value;` at end-of-file. The import
is now merged into the existing block; rustfmt could not parse the file to
notice it before.

curator 5/5 and media_resilience 34/34 on PostgreSQL, media_health 5/11 -> 9/11.

+++ crates/db/src/governance.rs
@@ -22,7 +22,8 @@ async fn fetch_level_postgres(
     account: &str,
 ) -> Result<i64, sqlx::Error> {
     let row = sqlx::query(sql).bind(account).fetch_optional(pool).await?;
-    Ok(row.map(|r| r.get::<i32, _>(0) as i64).unwrap_or(TL_NEW))
+    // COUNT() is INT8 on PostgreSQL, so the column is i64, not i32.
+    Ok(row.map(|r| r.get::<i64, _>(0)).unwrap_or(TL_NEW))
 }
 
 pub async fn trust_for(db: &Database, account: &str) -> Result<i64, sqlx::Error> {
